Refactor nzbToMediaSceneExceptions -> scene_exceptions
