+++ core/auto_process/tv.py
@@ -11,7 +11,7 @@ import requests
 import core
 from core import logger, transcoder
 from core.forks import auto_fork
-from core.nzbToMediaSceneExceptions import process_all_exceptions
+from core.scene_exceptions import process_all_exceptions
 from core.nzbToMediaUtil import convert_to_ascii, flatten, import_subs, list_media_files, remote_dir, report_nzb, remove_dir, server_responding
 
 requests.packages.urllib3.disable_warnings()
