+++ core/scene_exceptions.py
